improvement!: rename LSP3UniversalProfile -> UniversalProfile (#119)
* improvement!: rename LSP3UniversalProfile -> UniversalProfile

* docs: put arrays of addresses on one line

* chore: remove references to lsp3UniversalProfile
