+++ README.md
@@ -2,7 +2,7 @@
 
 <p align="center">
  <h2 align="center"><strong>@lukso/lsp-factory.js</strong></h2>
- <p align="center">Helper library to allow simple deployments of <a href="https://github.com/lukso-network/LIPs/blob/main/LSPs/LSP-0-ERC725Account.md">LSP3UniversalProfiles</a> and <a href="https://github.com/lukso-network/LIPs/blob/main/LSPs/LSP-4-DigitalCertificate.md">LSP7</a> and <a href="https://github.com/lukso-network/LIPs/blob/main/LSPs/LSP-8-IdentifiableDigitalAsset.md">LSP8 </a>Digital Assets.</p>
+ <p align="center">Helper library to allow simple deployments of <a href="https://github.com/lukso-network/LIPs/blob/main/LSPs/LSP-0-ERC725Account.md">UniversalProfiles</a> and <a href="https://github.com/lukso-network/LIPs/blob/main/LSPs/LSP-4-DigitalCertificate.md">LSP7</a> and <a href="https://github.com/lukso-network/LIPs/blob/main/LSPs/LSP-8-IdentifiableDigitalAsset.md">LSP8 </a>Digital Assets.</p>
 </p>
 
 <p align="center">For more information see <a href="https://docs.lukso.tech/tools/lsp-factoryjs/getting-started">Documentation</a>.</p>
@@ -32,40 +32,40 @@ const lspFactory = new LSPFactory(provider, {
 
 ```javascript
 // Deploy LSP3 Account
-const myContracts = await lspFactory.LSP3UniversalProfile.deploy({
-    controllerAddresses: ['0x...'], // Address which will controll the UP
-    lsp3Profile: {
-      json: {
-        LSP3Profile: {
-          name: "My Universal Profile",
-          description: "My cool Universal Profile",
-          profileImage: [
-            {
-              width: 500,
-              height: 500,
-              hashFunction: "keccak256(bytes)",
-              // bytes32 hex string of the image hash
-              hash: "0xfdafad027ecfe57eb4ad047b938805d1dec209d6e9f960fc320d7b9b11cbed14",
-              url: "ipfs://QmPLqMFHxiUgYAom3Zg4SiwoxDaFcZpHXpCmiDzxrtjSGp",
-            },
-          ],
-          backgroundImage: [
-            {
-              width: 500,
-              height: 500,
-              hashFunction: "keccak256(bytes)",
-              // bytes32 hex string of the image hash
-              hash: "0xfdafad027ecfe57eb4ad047b938805d1dec209d6e9f960fc320d7b9b11cbed14",
-              url: "ipfs://QmPLqMFHxiUgYAom3Zg4SiwoxDaFcZpHXpCmiDzxrtjSGp",
-            },
-          ],
-          tags: ['Fashion', 'Design'],
-          links: [{ title: "My Website", url: "www.my-website.com" }],
-        },
+const myContracts = await lspFactory.UniversalProfile.deploy({
+  controllerAddresses: ['0x...'], // Address which will controll the UP
+  lsp3Profile: {
+    json: {
+      LSP3Profile: {
+        name: 'My Universal Profile',
+        description: 'My cool Universal Profile',
+        profileImage: [
+          {
+            width: 500,
+            height: 500,
+            hashFunction: 'keccak256(bytes)',
+            // bytes32 hex string of the image hash
+            hash: '0xfdafad027ecfe57eb4ad047b938805d1dec209d6e9f960fc320d7b9b11cbed14',
+            url: 'ipfs://QmPLqMFHxiUgYAom3Zg4SiwoxDaFcZpHXpCmiDzxrtjSGp',
+          },
+        ],
+        backgroundImage: [
+          {
+            width: 500,
+            height: 500,
+            hashFunction: 'keccak256(bytes)',
+            // bytes32 hex string of the image hash
+            hash: '0xfdafad027ecfe57eb4ad047b938805d1dec209d6e9f960fc320d7b9b11cbed14',
+            url: 'ipfs://QmPLqMFHxiUgYAom3Zg4SiwoxDaFcZpHXpCmiDzxrtjSGp',
+          },
+        ],
+        tags: ['Fashion', 'Design'],
+        links: [{ title: 'My Website', url: 'www.my-website.com' }],
       },
-      url: "",
     },
-  });
+    url: '',
+  },
+});
 
 const myUPAddress = myContracts.ERC725Account.address;
 ```
@@ -77,8 +77,8 @@ The `deployReactive` option can be used to for real-time frontend updates.
 ```javascript
 const profileDeploymentEvents = [];
 
-const myUPAccount = await lspFactory.LSP3UniversalProfile.deploy({
-    deployReactive: true
+const myUPAccount = await lspFactory.UniversalProfile.deploy({
+  deployReactive: true,
 }).subscribe({
   next: (deploymentEvent: DeploymentEvent) => {
     profileDeploymentEvents.push(deploymentEvent);
