+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@lukso/lsp-factory.js",
   "version": "2.2.0",
-  "description": "Helper Library to allow simple deployments of LSP3UniversalProfiles and LSP7 and LSP8 Digital Assets.",
+  "description": "Helper Library to allow simple deployments of UniversalProfiles and LSP7 and LSP8 Digital Assets.",
   "main": "build/main/src/index.js",
   "typings": "build/main/src/index.d.ts",
   "module": "build/module/src/index.js",
