+++ src/index.ts
@@ -1,5 +1,5 @@
 export { LSPFactory } from './lib/lsp-factory';
-export { LSP3UniversalProfile } from './lib/classes/lsp3-universal-profile';
+export { UniversalProfile } from './lib/classes/universal-profile';
 export { LSP4DigitalAssetMetadata } from './lib/classes/lsp4-digital-asset-metadata';
 export { LSP8IdentifiableDigitalAsset } from './lib/classes/lsp8-identifiable-digital-asset';
 export { LSP7DigitalAsset } from './lib/classes/lsp7-digital-asset';
